Complete Merge

+++ LastClass/HorizontalScrollBar.cpp
@@ -1,3 +1,5 @@
+//HorizontalScrollBar.cpp
+
 #include "HorizontalScrollBar.h"
 #include "ClassDiagramForm.h"
 
@@ -28,10 +30,10 @@ void HorizontalScrollBar::OnHScrollLineRight() {
 	SCROLLINFO vScrinfo;
 	this->scrollBar->GetScrollInfo(&vScrinfo);
 
-	if ((vScrinfo.nPos + vScrinfo.nPage) < vScrinfo.nMax) {
+	if ((vScrinfo.nPos + (int)vScrinfo.nPage) < vScrinfo.nMax) {
 		Long height = 50;
 		vScrinfo.nPos += height;
-		if (vScrinfo.nPos + vScrinfo.nPage > vScrinfo.nMax) {
+		if (vScrinfo.nPos + (int)vScrinfo.nPage > vScrinfo.nMax) {
 			vScrinfo.nPos = vScrinfo.nMax - height;
 		}
 	}
